添加自动注册流程 & 将 register 合并进 larkuser 插件
