+++ src/plugins/nonebot_plugin_larkuser/utils/matcher.py
@@ -1,13 +1,18 @@
 from nonebot.matcher import Matcher
+from nonebot_plugin_orm import get_session
+from nonebot_plugin_userinfo import UserInfo, EventUserInfo
 
 from nonebot_plugin_larkutils import get_user_id
+from .register import register_user
 from ..lang import lang
 from .user import get_user
 
 
-async def check_access(user_id: str = get_user_id()) -> None:
+async def check_access(user_id: str = get_user_id(), user_info: UserInfo = EventUserInfo()) -> None:
     if (await get_user(user_id)).register_time is None:
-        await lang.finish("matcher.not_registered", user_id)
+        await lang.send("matcher.not_registered", user_id)
+        async with get_session() as session:
+            await register_user(session, user_id, user_info)
 
 
 def patch_matcher(matcher: type[Matcher]) -> type[Matcher]:
