--- src/plugins/nonebot_plugin_registry/lang.py
@@ -1,3 +1,0 @@
-from nonebot_plugin_larklang import LangHelper
-
-lang = LangHelper()
